RAG/requirements.txt
snowflake-snowpark-python
flask
snowflake
langchain
pandas
PyPDF2
docx2txt
python-docx